+++ ui/vue.config.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+//const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
 process.env.VUE_APP_SERVER_PORT = require('../test/homebridge/config.json').platforms[0].port || 8181;
 
@@ -139,7 +139,6 @@ module.exports = {
         'jquery.ui.touch-punch': path.resolve(__dirname, 'node_modules/gridstack/dist/jq/jquery.ui.touch-punch.js'),
       },
     },
-    plugins: process.env.NODE_ENV === 'production' ? [] : [new BundleAnalyzerPlugin()],
     //plugins: [new BundleAnalyzerPlugin()],
   },
 };